fix: adapt to rusqlite 0.40 and clippy 1.97
rusqlite no longer implements FromSql/ToSql for usize, so the two
affected sites go through i64 explicitly. Also resolves the new
useless_borrows_in_formatting lint (three sites) so the clippy gate
stays clean on current stable.

+++ crates/reprise-gnome/src/ui/device_sync/device_sync_planned_tests.rs
@@ -439,7 +439,7 @@ fn select_road_playlist(conn: &Rc<RefCell<Connection>>, ids: &[i64]) {
         conn.borrow()
             .execute(
                 "INSERT INTO playlist_tracks (playlist_id, track_id, position) VALUES (10, ?1, ?2)",
-                rusqlite::params![track_id, position],
+                rusqlite::params![track_id, position as i64],
             )
             .unwrap();
     }
